user,diary,comment 기본 구조
