+++ server/src/main/java/com/seb42/main30/seb42_main_030/diary/controller/DiaryController.java
@@ -1,14 +1,8 @@
-package com.seb42.main30.seb42_main_030.domain.diary.controller;
+package com.seb42.main30.seb42_main_030.diary.controller;
 
-import com.seb42.main30.seb42_main_030.domain.diary.dto.DiaryPostDto;
-import com.seb42.main30.seb42_main_030.domain.diary.entity.Diary;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
 
 @RestController
 @RequiredArgsConstructor
