+++ server/src/main/java/com/seb42/main30/seb42_main_030/diary/entity/Diary.java
@@ -1,4 +1,4 @@
-package com.seb42.main30.seb42_main_030.domain.diary.entity;
+package com.seb42.main30.seb42_main_030.diary.entity;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
