+++ server/src/main/java/com/seb42/main30/seb42_main_030/diary/service/DiaryService.java
@@ -1,0 +1,5 @@
+package com.seb42.main30.seb42_main_030.diary.service;
+
+
+public class DiaryService {
+}
