--- server/src/main/java/com/seb42/main30/seb42_main_030/domain/diary/service/DiaryService.java
@@ -1,5 +1,0 @@
-package com.seb42.main30.seb42_main_030.domain.diary.service;
-
-
-public class DiaryService {
-}
