--- server/src/main/java/com/seb42/main30/seb42_main_030/domain/user/repository/UserRepository.java
@@ -1,4 +1,0 @@
-package com.seb42.main30.seb42_main_030.domain.user.repository;
-
-public class UserRepository {
-}
