+++ server/src/main/java/com/seb42/main30/seb42_main_030/user/repository/UserRepository.java
@@ -1,0 +1,4 @@
+package com.seb42.main30.seb42_main_030.user.repository;
+
+public class UserRepository {
+}
